refactor(settings): split 1745-line settings component into 11 focused sub-components
Each sub-component is self-contained with OnPush change detection, its own collapsed
state, and handles its own service interactions. Parent is reduced to ~100 lines.
